审核员、审核管理员、收单员框架

+++ src/components/Admin/usermanage/useradd.vue
@@ -1,17 +1,38 @@
 <template>
   <div id="useradd">
-    <el-row >
+    <addItem :batchdelable="batchdelable" @adduser="hanldeAdd" @BatchIn="BatchIn" @BatchDel="BatchDel"></addItem>
+    <!-- <el-row >
       <el-button type="primary" size="small" icon="el-icon-edit-outline" @click="hanldeAdd()">添加用户</el-button>
       <el-button type="success" size="small" icon="el-icon-upload" @click="BatchIn()">批量导入</el-button>
-    </el-row>
+    </el-row> -->
+    <!-- <div class="search_container searchArea">
+        <el-form 
+            :inline="true" 
+            class="demo-form-inline search-form">
+            <el-form-item>
+                <el-button type="primary" size ="mini" icon="search" @click='adduser()'>添加用户</el-button>
+            </el-form-item>
+            <el-form-item class="btnRight">
+                <el-button type="success" size ="mini" icon="view" @click='BatchIn()'>批量导入</el-button>
+                <el-button type="danger" size ="mini" icon="view" @click='BatchDel()' :disabled="batchdelable" >批量删除</el-button>
+            </el-form-item>
+        </el-form>
+    </div> -->
     <template>
       <el-table
         :data="userData"
         stripe
-        style="width: 100%">
+        style="width: 100%"
+        @select="handleSelectionChange"
+        @select-all="selectAll">
+        <el-table-column
+          type="selection"
+          align='center'
+          width="60">
+        </el-table-column>
         <el-table-column
           type="index"
-          :index="indexMethod">
+          width="20">
         </el-table-column>
         <el-table-column
           label="编号"
@@ -106,7 +127,7 @@
           </template>
         </el-table-column>
       </el-table>
-      <el-pagination background layout="total"  :total="pageInfo.pageTotal">
+      <el-pagination background layout="total"  :total="pageTotal">
       </el-pagination>
     </template>
     <UserInfo :dialogAdd="dialogAdd" @update="getUserInfo"></UserInfo>
@@ -120,6 +141,7 @@
   import UserInfo from './UserInfo.vue'
   import UserEdit from './UserEdit.vue'
   import BatchIn from './BatchIn.vue'
+  import addItem from './addItem.vue'
 
   export default {
     name:'useradd',
@@ -127,7 +149,8 @@
     components:{   
       UserInfo,
       UserEdit,
-      BatchIn
+      BatchIn,
+      addItem
     },
 
     data() {
@@ -138,14 +161,14 @@
         dialogEdit : {   //编辑弹出框，默认是false
           show : false
         },
+        batchdelable : true ,
         dialogBatchIn:{
           show : false
         },
-        pageInfo: { //分页
-                pageTotal: 1
-            },
+        pageTotal: 0,
+        searchBtnDisabled : true ,
         form:{    //编辑信息
-          id:'',
+          /* id:'',
           username:'',
           depart:'',
           age:'',
@@ -154,22 +177,28 @@
           phone:'',
           salaryid:'',
           address:'',
-          role:''
+          role:'' */
         },
         userData:[{
-          id:'131231231',
-          username:'1231313',
-          depart:'13123123',
+          id:'131',
+          username:'123',
+          depart:'131',
           age:'',
           sex:'',
           nation:'',
           phone:'',
           salaryid:'',
           address:'',
-          role:'12313123'
-        }]
+          role:'123'
+        }],
+        multipleSelection: []
       }
     },
+    computed:{
+      newbatchdelable() {
+        return this.batchdelable;
+      }
+      },
     created(){
             this.getUserInfo();
         },
@@ -183,27 +212,50 @@
       getUserInfo() {
         this.$axios.get('http://localhost:3000/data').then(res => {
           this.pageInfo.pageTotal=res.count;
-          this.userData = res.data  
+          this.userData = res.data.userdate;
+          this.total = res.date.total
           //this.userData.push(formDate);
         })
       },
     
       handleDelete(index,row) {
         // 删除用户信息
-        this.$axios.delete(`http://localhost:3000/data/${row.id}`).then(res =>{
-            this.$message({
-                res,
-                type:"success",
-                message:"删除信息成功"
-            })
-            this.getUserInfo()    //删除数据，更新视图
+        this.$confirm('确认删除该用户吗?', '提示', {
+            type: 'warning'
+        })
+        .then(() => {
+          this.$axios.delete(`http://localhost:3000/data/${row.id}`).then(res =>{
+              this.$message({
+                  res,
+                  type:"success",
+                  message:"删除信息成功"
+              })
+              this.getUserInfo()    //删除数据，更新视图
+          })
         })
       },
+      BatchDel(){
+        this.$confirm('确认批量删除记录吗?', '提示', {
+            type: 'warning'
+        })
+        .then(() => {
+            const ids = this.rowIds.map(item => item.id).toString()
+            const para = { ids: ids }
+            Request(para).then(res => {
+                this.$message({
+                    message: '批量删除成功',
+                    type: 'success'
+                })
+                this.getUserInfo()
+            })
+        })
+        .catch(() => {})
+      },
+
       handleEdit(index,row){  //编辑
         this.dialogEdit.show = true ;  //显示弹
-        
-        this.form = {
-          id:row.id,
+        this.form = {...row
+          /* id:row.id,
           username:row.username,
           depart:row.depart,
           age:row.age,
@@ -212,13 +264,42 @@
           phone:row.phone,
           salaryid:row.salaryid,
           address:row.address,
-          role:row.role
+          role:row.role */
         }
-        console.log('adadad')
       },
+      handleSelectionChange(val) { 
+        
+        this.setbatchdelable(val);
+      },
+      selectAll(val){
+        val.forEach((item) => {
+          //this.rowIds.push(item.id); 
+        });
+        this.setbatchdelable(val);
+      },
+      setbatchdelable(val){
+        this.batchdelableg = true;
+        if(val.length > 0){
+          this.batchdelable = false;
+        }else{
+          this.batchdelable = true;
+        }
+      }
     }
   }
 </script>
 
 <style lang = "less" scoped>
+/* .search_container{
+        margin-bottom: 20px;
+    }
+    .btnRight{
+        float: right;
+        margin-right: 0px !important;
+    }
+    .searchArea{
+        background:rgba(255,255,255,1);
+        border-radius:2px;
+        padding: 18px 18px 0;
+    } */
 </style>
